pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Syed Muhammad Dawoud Sheraz Ali", email = "dawoud.sheraz@gmail.com"},
]
dependencies = [
    "factory-boy>=3.3.0",
    "nplusone>=1.0.0",
    "django-debug-toolbar>=4.1.0",
    "django-autoslug>=1.9.9",
    "djangorestframework>=3.14.0",
    "django>=4.2.4",
]
requires-python = ">=3.9"
license = {text = "MIT"}
